Create Chain
This code defines three classes: Block, Transaction, and Blockchain. The Block class represents a block in the blockchain and has a timestamp, a list of transactions, a hash of the previous block, and a hash of the current block (which is calculated based on the timestamp, transactions, and previous hash). The Transaction class represents a single transaction of energy from one user to another and has a sender, a receiver, and an amount of energy.

The Blockchain class represents the entire blockchain and has a list of blocks and a list of pending transactions. Users can add transactions to the blockchain using the add_transaction method, which adds the transaction to the list of pending transactions. Users can then "mine" a block using the mine_block method, which creates a new block with the pending transactions and adds it to the chain. Finally, users can print out the entire blockchain using the print_chain method.

Again, this is a very basic outline and doesn't include features such as proof of work or consensus algorithms. However, it should give you an idea of how a peer-to-peer energy trading system could be implemented using a simple blockchain in Python.